+++ attributes/default.rb
@@ -1,2 +1,3 @@
 default['zabbix']['version'] = '2.4'
 default['zabbix']['api-version'] = '2.4.7'
+default['zabbix']['db_vendor'] = 'postgresql'